Moved data file to data dir
